Refactor display.py as module & Add menu to MainWindow
