+++ laser_setup/display/experiment_window.py
@@ -1,92 +1,21 @@
-"""
-This module contains the display functions for the GUI.
-"""
-import os
-import sys
 import time
 import logging
-from importlib.metadata import metadata
 from typing import Type
 
 from pymeasure.experiment import unique_filename, Results, Procedure
-from pymeasure.display.widgets import InputsWidget, TabWidget, PlotFrame
+from pymeasure.display.widgets import InputsWidget, PlotFrame
 from pymeasure.display.windows import ManagedWindow
 from pymeasure.display.Qt import QtGui, QtWidgets, QtCore
 
-from . import config, config_path, _config_file_used
-from .utils import remove_empty_data
-from .procedures import MetaProcedure, BaseProcedure, ChipProcedure
+from .. import config
+from .widgets import TextWidget, ProgressBar
+from ..procedures import MetaProcedure, BaseProcedure, ChipProcedure
 
 log = logging.getLogger(__name__)
 
 
-class ProgressBar(QtWidgets.QDialog):
-    """A simple progress bar dialog."""
-    def __init__(self, parent=None, title="Waiting", text=""):
-        super().__init__(parent)
-        self.setWindowTitle(title)
-        self._layout = QtWidgets.QVBoxLayout(self)
-        self.label = QtWidgets.QLabel(self)
-        self.label.setText(text)
-        self.progress = QtWidgets.QProgressBar(self)
-        self._layout.addWidget(self.label)
-        self._layout.addWidget(self.progress)
-        self.setLayout(self._layout)
-        self.timer = QtCore.QTimer(self)
-        self.timer.timeout.connect(self._update_progress)
-
-    def start(self, wait_time: float, fps: float = 30., decimals: int = 0):
-        self.wait_time = wait_time
-        self.frame_interval = 1 / fps
-        self.total_frames = int(fps * wait_time)
-        self.start_time = time.perf_counter()
-        self.progress.setRange(0, self.total_frames)
-        self.show()
-        self.timer.start(max(1, round(1000 / fps)))
-        self.d = decimals
-
-    def _update_progress(self):
-        elapsed_time = time.perf_counter() - self.start_time
-        current_frame = int(elapsed_time / self.frame_interval)
-
-        if current_frame >= self.total_frames:
-            self.progress.setValue(self.total_frames)
-            self.progress.setFormat(
-                f"{self.wait_time:.{self.d}f} / {self.wait_time:.{self.d}f} s"
-            )
-            self.timer.stop()
-            self.close()
-        else:
-            self.progress.setValue(current_frame)
-            self.progress.setFormat(
-                f"{elapsed_time:.{self.d}f} / {self.wait_time:.{self.d}f} s"
-            )
-
-
-class TextWidget(TabWidget, QtWidgets.QWidget):
-    def __init__(self, name: str = None, parent=None, file: str = None):
-        super().__init__(name, parent)
-        self.view = QtWidgets.QTextEdit(self, readOnly=True)
-        self.view.setReadOnly(True)
-        self.view.setStyleSheet("""
-            font-size: 12pt;
-        """)
-        try:
-            with open(file, encoding='utf-8') as f:
-                readme_text = f.read()
-        except:
-            readme_text = f'{file} not found :('
-        self.view.setMarkdown(readme_text)
-
-        vbox = QtWidgets.QVBoxLayout(self)
-        vbox.setSpacing(0)
-
-        vbox.addWidget(self.view)
-        self.setLayout(vbox)
-
-
 class ExperimentWindow(ManagedWindow):
-    """The main window for the GUI. It is used to display a
+    """The main window for an experiment. It is used to display a
     `pymeasure.experiment.Procedure`, and allows for the experiment to be run
     from the GUI, by queuing it in the manager. It also allows for existing
     data to be loaded and displayed.
@@ -355,204 +284,3 @@ class MetaProcedureWindow(QtWidgets.QMainWindow):
 
         else:
             time.sleep(wait_time)
-
-
-class MainWindow(QtWidgets.QMainWindow):
-    def __init__(
-            self,
-            sequences: dict[str, Type[MetaProcedure]],
-            experiments: dict[str, Type[Procedure]],
-            scripts: dict[str, callable],
-            **kwargs
-        ):
-        super().__init__(**kwargs)
-        self.setWindowTitle('Laser Setup')
-        self.setWindowIcon(
-            self.style().standardIcon(QtWidgets.QStyle.StandardPixmap.SP_ComputerIcon)
-        )
-        self.resize(640, 480)
-        self.setCentralWidget(QtWidgets.QWidget())
-
-        self.sequences = sequences
-        self.experiments = experiments
-        self.scripts = scripts
-        self.gridx = max(len(experiments), len(scripts), 3)
-        self.windows = {}
-
-        # Experiment Buttons
-        self._layout = QtWidgets.QGridLayout(self.centralWidget())
-        self.buttons = {}
-        for i, (cls, name) in enumerate(experiments.items()):
-            self.buttons[cls] = QtWidgets.QPushButton(name)
-            self.buttons[cls].clicked.connect(self.open_app(cls))
-            self.buttons[cls].setToolTip(cls.__doc__)
-            self._layout.addWidget(self.buttons[cls], 2, i)
-
-        # README Widget
-        readme = QtWidgets.QTextEdit(readOnly=True)
-        readme.setStyleSheet("""
-            font-size: 12pt;
-        """)
-        try:
-            with open('README.md') as f:
-                readme_text = f.read()
-        except FileNotFoundError:
-            readme_text = metadata('laser_setup').get('Description')
-        readme.setMarkdown(readme_text)
-        self._layout.addWidget(readme, 1, 0, 1, self.gridx)
-
-        # Settings Button
-        settings = QtWidgets.QPushButton('Settings')
-        settings.clicked.connect(self.edit_settings)
-        self._layout.addWidget(settings, 0, 0)
-
-        # Secuences Buttons
-        for i, (cls, name) in enumerate(sequences.items()):
-            self.buttons[cls] = QtWidgets.QPushButton(name)
-            self.buttons[cls].clicked.connect(self.open_sequence(cls))
-            self.buttons[cls].setToolTip(cls.__doc__)
-            self._layout.addWidget(self.buttons[cls], 0, 1+i)
-
-        # Reload window button
-        self.reload = QtWidgets.QPushButton('Reload')
-        self.reload.clicked.connect(
-            lambda: os.execl(sys.executable, sys.executable, '.', *sys.argv[1:])
-        )
-        self._layout.addWidget(self.reload, 0, self.gridx-1)
-
-        for i, (func, name) in enumerate(scripts.items()):
-            self.buttons[func] = QtWidgets.QPushButton(name)
-            self.buttons[func].clicked.connect(self.run_script(func))
-            self.buttons[func].setToolTip(func.__doc__)
-            self._layout.addWidget(self.buttons[func], 3, i)
-
-    def open_sequence(self, cls: Type[MetaProcedure]):
-        def func():
-            self.windows[cls] = MetaProcedureWindow(cls, title=self.sequences[cls], parent=self)
-            self.windows[cls].show()
-            self.suggest_reload()
-        return func
-
-    def open_app(self, cls: Type[Procedure]):
-        def func():
-            self.windows[cls] = ExperimentWindow(cls, title=self.experiments[cls], parent=self)
-            self.windows[cls].show()
-        return func
-
-    def run_script(self, f: callable):
-        def func():
-            try:
-                f(parent=self)
-            except TypeError:
-                f()
-            self.suggest_reload()
-        return func
-
-    def edit_settings(self):
-        if _config_file_used != config_path:
-            choice = self.select_from_list('No config file found',
-                ['Create new config', 'Use default config'], 'Select an option:')
-
-            if choice == 'Create new config':
-                os.makedirs(os.path.dirname(config_path), exist_ok=True)
-                with open(config_path, 'w') as f:
-                    config.write(f)
-
-            elif choice == 'Use default config':
-                return
-
-        os.startfile(config_path.replace('/', '\\'))
-        self.suggest_reload()
-
-    def suggest_reload(self):
-        self.reload.setStyleSheet('background-color: red;')
-        self.reload.setText('Reload to apply changes')
-
-    def error_dialog(self, message:str):
-        error_dialog = QtWidgets.QMessageBox()
-        error_dialog.setWindowTitle("Error")
-        error_dialog.setText(f"An error occurred:\n{message}\nPlease reload the program.")
-        error_dialog.setIcon(QtWidgets.QMessageBox.Icon.Critical)
-        error_dialog.exec()
-        self.reload.click()
-
-    def lock_window(self, msg: str = ''):
-        self.setEnabled(False)
-        self.lock_msg = msg
-        self.locked = QtWidgets.QMessageBox(QtWidgets.QMessageBox.Icon.Information, 'Locked', msg)
-        self.locked.exec()
-
-    def select_from_list(self, title: str, items: list[str], label: str = '') -> str:
-        item, ok = QtWidgets.QInputDialog.getItem(self, title, label, items, 0, False)
-        if ok:
-            return item
-        return None
-
-    def question_box(self, title: str, text: str) -> bool:
-        buttons = QtWidgets.QMessageBox.StandardButton.Yes | QtWidgets.QMessageBox.StandardButton.No
-        reply = QtWidgets.QMessageBox.question(self, title, text, buttons)
-        return reply == QtWidgets.QMessageBox.StandardButton.Yes
-
-
-def display_window(Window: Type[QtWidgets.QMainWindow], *args, **kwargs):
-    """Displays the window for the given class. Allows for the
-    window to be run from the GUI, by queuing it in the manager.
-    It also allows for existing data to be loaded and displayed.
-
-    :param Window: The Qt Window subclass to display.
-    :param args: The arguments to pass to the window class.
-    """
-    app = QtWidgets.QApplication(sys.argv)
-    app.setStyle(config['GUI']['style'])    # Get available styles with QtWidgets.QStyleFactory.keys()
-    if bool(eval(config['GUI']['dark_mode'])):
-        app.setPalette(get_dark_palette())
-    QtCore.QLocale.setDefault(QtCore.QLocale(
-        QtCore.QLocale.Language.English,
-        QtCore.QLocale.Country.UnitedStates
-    ))
-    window = Window(*args, **kwargs)
-    window.show()
-    app.exec()
-    remove_empty_data()
-    sys.exit()
-
-
-def display_experiment(cls: Type[Procedure], title: str = ''):
-    """Wrapper around display_window for ExperimentWindow.
-    TODO: Remove this function and use display_window directly.
-    """
-    display_window(ExperimentWindow, cls, title)
-
-
-def get_dark_palette():
-    palette = QtGui.QPalette()
-    ColorRole = QtGui.QPalette.ColorRole
-
-    # Set the background color
-    palette.setColor(ColorRole.Window, QtGui.QColor(50, 50, 50))
-
-    # Set the title text color
-    palette.setColor(ColorRole.WindowText, QtGui.QColor(200, 200, 200))
-
-    # Set the input text color
-    palette.setColor(ColorRole.Text, QtGui.QColor(200, 200, 200))
-
-    # Set the button color
-    palette.setColor(ColorRole.Button, QtGui.QColor(30, 30, 30))
-
-    # Set the button text color
-    palette.setColor(ColorRole.ButtonText, QtGui.QColor(200, 200, 200))
-
-    # Set the base color
-    palette.setColor(ColorRole.Base, QtGui.QColor(35, 35, 35))
-
-    # Set the alternate background color
-    palette.setColor(ColorRole.AlternateBase, QtGui.QColor(45, 45, 45))
-
-    # Set the highlight color
-    palette.setColor(ColorRole.Highlight, QtGui.QColor(42, 130, 218))
-
-    # Set the highlighted text color
-    palette.setColor(ColorRole.HighlightedText, QtGui.QColor(240, 240, 240))
-
-    return palette
